requirement.txt
appdirs==1.4.3
certifi==2017.4.17
chardet==3.0.3
Django==1.11.1
django-filter==1.0.4
djangorestframework==3.6.3
httpie==0.9.9
idna==2.5
Markdown==2.6.8
packaging==16.8
Pygments==2.2.0
pyparsing==2.2.0
pytz==2017.2
requests==2.17.3
six==1.10.0
urllib3==1.21.1
